pyproject.toml
[project]
name = "paper-radar"
version = "0.1.0"
description = "PaperRadar - Automated paper analysis based on keywords using dual LLM architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.1.3",
    "feedparser>=6.0.11",
    "fastapi>=0.115.0",
    "openai>=1.57.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.2",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.1",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
